fix popup information edge cases
