+++ assignment/assignment1/client/src/scenes/widgets/QuestionFields/ConstraintsFields.tsx
@@ -1,8 +1,6 @@
 import { useState } from "react";
-import { Button,TextField} from "@mui/material";
-import {DeleteOutlined} from "@mui/icons-material";
-import AddCircleOutlinedIcon from '@mui/icons-material/AddCircleOutlined';
 import { Question } from "../../../state/question";
+import { ConstraintsFieldsUI } from "./QuestionFieldsFormUI";
 
 interface ConstraintsFileds {
     newData: Partial<Question>; 
@@ -11,24 +9,15 @@ interface ConstraintsFileds {
 
 const ConstraintsFields: React.FC<ConstraintsFileds> =({newData, setData}) => {
     const [fields, setFields] = useState([""]);
-    const maxInputNum = 4;
-
-    const TextFieldCSS = {
-        width: "80%",
-        mb: 1, // mb= margin-bottom
-    };
-
-    const ButtonIconCSS = {
-        ml:1,
-        maxWidth: '30px', 
-        maxHeight: '30px', 
-        minWidth: '30px', 
-        minHeight: '30px'
-    }
 
     const updateData = (fields : any) => {
         setData({ ...newData, constraints: fields})
     }
+
+    const handleUpdateField = (updatedFields:any) => {
+        setFields(updatedFields);
+        updateData(updatedFields);
+    };
 
     const handleAddField = () => {
         setFields([...fields, ""]);
@@ -36,40 +25,18 @@ const ConstraintsFields: React.FC<ConstraintsFileds> =({newData, setData}) => {
 
     const handleRemoveField = (index:any)=> {
         const filteredFields = fields.filter((_,i)=> i != index);
-        setFields(filteredFields);
-        updateData(filteredFields);
+        handleUpdateField(filteredFields);
     };
     
-    return (
-        <div>
-            {fields.map((field, index) => (
-                <div key={index}>
-                    <TextField
-                        sx={{...TextFieldCSS }}
-                        label="Constraints"
-                        placeholder="Enter Constraints"
-                        value={field}
-                        onChange={(e) => {
-                            const updatedFields = [...fields];
-                            updatedFields[index] = e.target.value;
-                            setFields(updatedFields);
-                            updateData(updatedFields);
-                        }}
-                    />
-                    {fields.length > 1 && (
-                        <Button sx={{...ButtonIconCSS}} onClick={() => handleRemoveField(index)}>
-                        <DeleteOutlined/>
-                        </Button>
-                    )}
-                    {fields.length - 1 === index && fields.length < maxInputNum && (
-                        <Button sx={{...ButtonIconCSS}} onClick={handleAddField}>
-                            <AddCircleOutlinedIcon/>
-                        </Button>
-                    )}
-                </div>
-            ))}
-        </div>
-      );
+    return(
+        <ConstraintsFieldsUI
+            fields={fields}
+            handleAddFields={handleAddField}
+            handleDeleteFields={handleRemoveField}
+            handleUpdateFields={handleUpdateField}
+        />
+    );
+    
 };
 
 export default ConstraintsFields;
