+++ assignment/assignment1/client/src/scenes/widgets/QuestionFields/ExampleFields.tsx
@@ -1,8 +1,6 @@
 import { useState } from "react";
-import { Button,Box, TextField} from "@mui/material";
-import {DeleteOutlined} from "@mui/icons-material";
-import AddCircleOutlinedIcon from '@mui/icons-material/AddCircleOutlined';
 import { Question } from "../../../state/question";
+import { ExampleFieldsUI } from "./QuestionFieldsFormUI";
 
 interface ExampleFileds {
     newData: Partial<Question>; 
@@ -15,19 +13,6 @@ const ExampleFields: React.FC<ExampleFileds> =({newData, setData}) => {
 	    outputText: "",
 	    explanation: ""
     }]);
-   
-    const TextFieldCSS = {
-        width: "80%",
-        mb: 1, // mb= margin-bottom
-    };
-
-    const ButtonIconCSS = {
-        ml:1,
-        maxWidth: '30px', 
-        maxHeight: '30px', 
-        minWidth: '30px', 
-        minHeight: '30px'
-    }
 
     const updateData = (fields : any) => {
         setData({ ...newData, examples: fields})
@@ -49,57 +34,12 @@ const ExampleFields: React.FC<ExampleFileds> =({newData, setData}) => {
     };
     
     return (
-        <div>
-            {fields.map((field, index) => (
-                <Box sx={{boxShadow: 1,}}>
-                    <div key={index}>
-                        <TextField
-                            sx={{...TextFieldCSS }}
-                            label="Example Input"
-                            placeholder="Enter Example Input"
-                            value={field.inputText}
-                            onChange={(e) => {
-                                const updatedFields = [...fields];
-                                updatedFields[index].inputText = e.target.value;
-                                handleUpdateField(updatedFields);
-                            }}
-                        />
-                        <TextField
-                            sx={{...TextFieldCSS }}
-                            label="Example Output"
-                            placeholder="Enter Example Output"
-                            value={field.outputText}
-                            onChange={(e) => {
-                                const updatedFields = [...fields];
-                                updatedFields[index].outputText = e.target.value;
-                                handleUpdateField(updatedFields);
-                            }}
-                        />
-                        <TextField
-                            sx={{...TextFieldCSS }}
-                            label="Example Explaination"
-                            placeholder="Enter Example Explaination"
-                            value={field.explanation}
-                            onChange={(e) => {
-                                const updatedFields = [...fields];
-                                updatedFields[index].explanation = e.target.value;
-                                handleUpdateField(updatedFields);
-                            }}
-                        />
-                        {fields.length > 1 && (
-                            <Button sx={{...ButtonIconCSS}} onClick={() => handleRemoveField(index)}>
-                            <DeleteOutlined/>
-                            </Button>
-                        )}
-                        {fields.length - 1 === index && fields.length < 3 && (
-                            <Button sx={{...ButtonIconCSS}} onClick={handleAddField}>
-                                <AddCircleOutlinedIcon/>
-                            </Button>
-                        )}
-                    </div>
-                </Box>
-            ))}
-        </div>
+        <ExampleFieldsUI
+            fields={fields}
+            handleUpdateFields={handleUpdateField}
+            handleAddFields={handleAddField}
+            handleDeleteFields={handleRemoveField}
+        />
       );
 };
